refactor(T045): extract shared grayscale helper (ImageProcessing.ToGrayscale) and move GrayImage to Domain.Vision; update ImageMatchEvaluator to reuse
